decouple threaded and hte tests; update deps
