+++ tests/test_convert.py
@@ -2,7 +2,7 @@
 Tests for CytoTable.convert and related.
 """
 
-# pylint: disable=no-member,too-many-lines
+# pylint: disable=no-member,too-many-lines,unused-argument
 
 import itertools
 import os
@@ -11,15 +11,10 @@ from shutil import copy
 from typing import Any, Dict, List, Tuple, cast
 
 import duckdb
-import parsl
 import pyarrow as pa
 import pyarrow.compute as pc
 import pytest
 from cloudpathlib import CloudPath
-from parsl.channels import LocalChannel
-from parsl.config import Config
-from parsl.executors import HighThroughputExecutor
-from parsl.providers import LocalProvider
 from pyarrow import csv, parquet
 from pycytominer.cyto_utils.cells import SingleCells
 
@@ -39,7 +34,6 @@ from cytotable.utils import (
     _column_sort,
     _duckdb_reader,
     _expand_path,
-    _parsl_loaded,
     _sqlite_mixed_type_query_to_parquet,
 )
 
@@ -86,7 +80,9 @@ def test_extend_path(fx_tempdir: str):
     assert _expand_path(path=f"{subdir}/..") == pathlib.Path(fx_tempdir).resolve()
 
 
-def test_get_source_filepaths(fx_tempdir: str, data_dir_cellprofiler: str):
+def test_get_source_filepaths(
+    load_parsl_default: None, fx_tempdir: str, data_dir_cellprofiler: str
+):
     """
     Tests _get_source_filepaths
     """
@@ -131,7 +127,7 @@ def test_get_source_filepaths(fx_tempdir: str, data_dir_cellprofiler: str):
     assert len(set(single_dir_result.keys())) == 4
 
 
-def test_prepend_column_name(fx_tempdir: str):
+def test_prepend_column_name(load_parsl_default: None, fx_tempdir: str):
     """
     Tests _prepend_column_name
     """
@@ -211,6 +207,7 @@ def test_prepend_column_name(fx_tempdir: str):
 
 
 def test_concat_source_group(
+    load_parsl_default: None,
     fx_tempdir: str,
     example_tables: Tuple[pa.Table, ...],
     example_local_sources: Dict[str, List[Dict[str, Any]]],
@@ -261,7 +258,7 @@ def test_concat_source_group(
         ).result()
 
 
-def test_get_join_chunks(fx_tempdir: str):
+def test_get_join_chunks(load_parsl_default: None, fx_tempdir: str):
     """
     Tests _get_join_chunks
     """
@@ -299,7 +296,7 @@ def test_get_join_chunks(fx_tempdir: str):
     ) == {"id1", "id2"}
 
 
-def test_join_source_chunk(fx_tempdir: str):
+def test_join_source_chunk(load_parsl_default: None, fx_tempdir: str):
     """
     Tests _get_join_chunks
     """
@@ -365,7 +362,7 @@ def test_join_source_chunk(fx_tempdir: str):
     )
 
 
-def test_concat_join_sources(fx_tempdir: str):
+def test_concat_join_sources(load_parsl_default: None, fx_tempdir: str):
     """
     Tests _concat_join_sources
     """
@@ -451,7 +448,9 @@ def test_concat_join_sources(fx_tempdir: str):
     )
 
 
-def test_infer_source_datatype():
+def test_infer_source_datatype(
+    load_parsl_default: None,
+):
     """
     Tests _infer_source_datatype
     """
@@ -474,7 +473,9 @@ def test_infer_source_datatype():
 
 
 def test_to_parquet(
-    fx_tempdir: str, example_local_sources: Dict[str, List[Dict[str, Any]]]
+    load_parsl_default: None,
+    fx_tempdir: str,
+    example_local_sources: Dict[str, List[Dict[str, Any]]],
 ):
     """
     Tests _to_parquet
@@ -526,133 +527,8 @@ def test_to_parquet(
         assert parquet_result.shape == csv_source.shape
 
 
-def test_convert_s3_path_csv(
-    fx_tempdir: str,
-    example_local_sources: Dict[str, List[Dict[str, Any]]],
-    example_s3_endpoint: str,
-):
-    """
-    Tests convert with mocked csv s3 object storage endpoint
-    """
-
-    multi_dir_nonconcat_s3_result = convert(
-        source_path="s3://example/",
-        dest_path=f"{fx_tempdir}/s3_test",
-        dest_datatype="parquet",
-        concat=False,
-        join=False,
-        joins=None,
-        source_datatype="csv",
-        compartments=["cytoplasm", "cells"],
-        metadata=["image"],
-        identifying_columns=["imagenumber"],
-        # endpoint_url here will be used with cloudpathlib client(**kwargs)
-        endpoint_url=example_s3_endpoint,
-    )
-
-    # compare each of the results using files from the source
-    for control_path, test_path in zip(
-        [
-            source["table"]
-            for group in cast(Dict, multi_dir_nonconcat_s3_result).values()
-            for source in group
-        ],
-        [
-            source["table"]
-            for group in example_local_sources.values()
-            for source in group
-        ],
-    ):
-        parquet_control = parquet.ParquetDataset(path_or_paths=control_path).read()
-        parquet_result = parquet.ParquetDataset(
-            path_or_paths=test_path, schema=parquet_control.schema
-        ).read()
-
-        assert parquet_result.schema.equals(parquet_control.schema)
-        assert parquet_result.shape == parquet_control.shape
-
-
-def test_convert_s3_path_sqlite(
-    fx_tempdir: str,
-    data_dir_cellprofiler_sqlite_nf1: str,
-    example_s3_endpoint: str,
-):
-    """
-    Tests convert with mocked sqlite s3 object storage endpoint
-
-    Note: we use a dedicated tmpdir for work in this test to avoid
-    race conditions with nested pytest fixture post-yield deletions.
-    """
-
-    # local sqlite read
-    local_cytotable_table = parquet.read_table(
-        source=convert(
-            source_path=data_dir_cellprofiler_sqlite_nf1,
-            dest_path=(
-                f"{fx_tempdir}/{pathlib.Path(data_dir_cellprofiler_sqlite_nf1).name}"
-                ".cytotable.parquet"
-            ),
-            dest_datatype="parquet",
-            chunk_size=100,
-            preset="cellprofiler_sqlite_pycytominer",
-        )
-    )
-
-    # s3 sqlite read with single and directly referenced file
-    s3_cytotable_table = parquet.read_table(
-        source=convert(
-            source_path=f"s3://example/nf1/{pathlib.Path(data_dir_cellprofiler_sqlite_nf1).name}",
-            dest_path=(
-                f"{fx_tempdir}/{pathlib.Path(data_dir_cellprofiler_sqlite_nf1).name}"
-                ".cytotable.parquet"
-            ),
-            dest_datatype="parquet",
-            chunk_size=100,
-            preset="cellprofiler_sqlite_pycytominer",
-            endpoint_url=example_s3_endpoint,
-            # use explicit cache to avoid temp cache removal / overlaps with
-            # sequential s3 SQLite files. See below for more information
-            # https://cloudpathlib.drivendata.org/stable/caching/#automatically
-            local_cache_dir=f"{fx_tempdir}/sqlite_s3_cache/1",
-        )
-    )
-
-    # s3 sqlite read with nested sqlite file
-    s3_cytotable_table_nested = parquet.read_table(
-        source=convert(
-            source_path="s3://example/nf1/",
-            dest_path=(
-                f"{fx_tempdir}/{pathlib.Path(data_dir_cellprofiler_sqlite_nf1).name}"
-                ".cytotable.parquet"
-            ),
-            dest_datatype="parquet",
-            chunk_size=100,
-            preset="cellprofiler_sqlite_pycytominer",
-            endpoint_url=example_s3_endpoint,
-            # use explicit cache to avoid temp cache removal / overlaps with
-            # sequential s3 SQLite files. See below for more information
-            # https://cloudpathlib.drivendata.org/stable/caching/#automatically
-            local_cache_dir=f"{fx_tempdir}/sqlite_s3_cache/2",
-        )
-    )
-
-    assert local_cytotable_table.sort_by(
-        [(name, "ascending") for name in local_cytotable_table.schema.names]
-    ).equals(
-        s3_cytotable_table.sort_by(
-            [(name, "ascending") for name in s3_cytotable_table.schema.names]
-        )
-    )
-    assert local_cytotable_table.sort_by(
-        [(name, "ascending") for name in local_cytotable_table.schema.names]
-    ).equals(
-        s3_cytotable_table_nested.sort_by(
-            [(name, "ascending") for name in s3_cytotable_table_nested.schema.names]
-        )
-    )
-
-
 def test_infer_source_group_common_schema(
+    load_parsl_default: None,
     example_local_sources: Dict[str, List[Dict[str, Any]]],
     example_tables: Tuple[pa.Table, ...],
 ):
@@ -669,6 +545,7 @@ def test_infer_source_group_common_schema(
 
 
 def test_convert_cytominerdatabase_csv(
+    load_parsl_default: None,
     fx_tempdir: str,
     data_dirs_cytominerdatabase: List[str],
     cytominerdatabase_to_pycytominer_merge_single_cells_parquet: List[str],
@@ -723,7 +600,10 @@ def test_convert_cytominerdatabase_csv(
 
 
 def test_convert_cellprofiler_sqlite(
-    fx_tempdir: str, data_dir_cellprofiler: str, cellprofiler_merged_nf1data: pa.Table
+    load_parsl_default: None,
+    fx_tempdir: str,
+    data_dir_cellprofiler: str,
+    cellprofiler_merged_nf1data: pa.Table,
 ):
     """
     Tests convert with cellprofiler sqlite exports
@@ -757,6 +637,7 @@ def test_convert_cellprofiler_sqlite(
 
 
 def test_convert_cellprofiler_csv(
+    load_parsl_default: None,
     fx_tempdir: str,
     data_dir_cellprofiler: str,
     cellprofiler_merged_examplehuman: pa.Table,
@@ -804,6 +685,7 @@ def test_convert_cellprofiler_csv(
 
 
 def test_cast_data_types(
+    load_parsl_default: None,
     fx_tempdir: str,
     data_dir_cellprofiler_sqlite_nf1: str,
 ):
@@ -901,6 +783,7 @@ def test_cast_data_types(
 
 
 def test_convert_cellprofiler_sqlite_pycytominer_merge(
+    load_parsl_default: None,
     fx_tempdir: str,
     data_dir_cellprofiler_sqlite_nf1: str,
 ):
@@ -976,7 +859,7 @@ def test_convert_cellprofiler_sqlite_pycytominer_merge(
 
 
 def test_sqlite_mixed_type_query_to_parquet(
-    fx_tempdir: str, example_sqlite_mixed_types_database: str
+    load_parsl_default: None, fx_tempdir: str, example_sqlite_mixed_types_database: str
 ):
     """
     Testing _sqlite_mixed_type_query_to_parquet
@@ -1053,70 +936,8 @@ def test_sqlite_mixed_type_query_to_parquet(
     }
 
 
-def test_convert_hte_cellprofiler_csv(
-    fx_tempdir: str,
-    data_dir_cellprofiler: str,
-    cellprofiler_merged_examplehuman: pa.Table,
-):
-    """
-    Tests convert with Parsl HighThroughputExecutor
-
-    See the following for more details.
-    https://parsl.readthedocs.io/en/stable/stubs/parsl.executors.HighThroughputExecutor.html#parsl.executors.HighThroughputExecutor
-    """
-
-    # check for previously loaded configuration
-    if _parsl_loaded():
-        # clear the previous config
-        parsl.clear()
-
-    local_htex = Config(
-        executors=[
-            HighThroughputExecutor(
-                label="htex_Local",
-                worker_debug=True,
-                cores_per_worker=1,
-                provider=LocalProvider(
-                    channel=LocalChannel(),
-                    init_blocks=1,
-                    max_blocks=1,
-                ),
-            )
-        ],
-        strategy=None,
-    )
-
-    control_result = cellprofiler_merged_examplehuman
-
-    test_result = parquet.read_table(
-        convert(
-            source_path=f"{data_dir_cellprofiler}/ExampleHuman",
-            dest_path=f"{fx_tempdir}/ExampleHuman",
-            dest_datatype="parquet",
-            source_datatype="csv",
-            preset="cellprofiler_csv",
-            # uses a custom parsl configuration
-            parsl_config=local_htex,
-        )
-    )
-
-    # sort all values by the same columns
-    # we do this due to the potential for inconsistently ordered results
-    control_result = control_result.sort_by(
-        [(colname, "ascending") for colname in control_result.column_names]
-    )
-    test_result = test_result.sort_by(
-        [(colname, "ascending") for colname in test_result.column_names]
-    )
-
-    assert test_result.shape == control_result.shape
-    assert test_result.equals(control_result)
-
-    # clean up the parsl config for other tests
-    parsl.clear()
-
-
 def test_cell_health_cellprofiler_to_cytominer_database_legacy(
+    load_parsl_default: None,
     fx_tempdir: str,
     data_dir_cytominerdatabase: str,
     fixture_cytominerdatabase_merged_cellhealth: pa.Table,
